+++ test/testproject/proxy/odatajs.d.ts
@@ -1,14 +1,12 @@
-declare module "odatajs" {
-    class oData {
+export class oData {
         static request(request: Request, success?: (data: any, response: any) => void, error?: (error: any) => void, handler?, httpClient?, metadata?);
     }
 
-    interface Request {
+export interface Request {
         requestUri: string,
         method: string,
         headers: Header | Header[],
         data?: any
     }
 
-    interface Header { [name: string]: string }
-}
+export interface Header { [name: string]: string }